+++ visualization/app/app.config.ts
@@ -22,6 +22,7 @@ import { UpdateQueryParametersEffect } from "app/codeCharta/state/effects/update
 import { UpdateVisibleTopLabelsEffect } from "app/codeCharta/state/effects/updateVisibleTopLabels/updateVisibleTopLabels.effect"
 import { ResetColorRangeEffect } from "app/codeCharta/state/store/dynamicSettings/colorRange/resetColorRange.effect"
 import { BlacklistSearchPatternEffect } from "app/codeCharta/ui/ribbonBar/searchPanel/searchBar/blacklistSearchPattern.effect"
+import { UpdateAmountOfEdgePreviewsEffect } from "./codeCharta/state/effects/amountOfEdgePreviews/updateAmountOfEdgePreviews.effect"
 
 export const appConfig: ApplicationConfig = {
     providers: [
@@ -33,6 +34,7 @@ export const appConfig: ApplicationConfig = {
         provideEffects([
             UnfocusNodesEffect,
             AddBlacklistItemsIfNotResultsInEmptyMapEffect,
+            UpdateAmountOfEdgePreviewsEffect,
             OpenNodeContextMenuEffect,
             BlacklistSearchPatternEffect,
             ResetColorRangeEffect,