+++ visualization/app/codeCharta/state/store/appSettings/showEdges/incoming/showIncomingEdges.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from "@ngrx/store"
+
+export const setShowIncomingEdges = createAction("SET_SHOW_INCOMING_EDGES", props<{ value: boolean }>())