+++ visualization/app/codeCharta/state/store/appSettings/showEdges/outgoing/showOutgoingEdges.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from "@ngrx/store"
+
+export const setShowOutgoingEdges = createAction("SET_SHOW_OUTGOING_EDGES", props<{ value: boolean }>())