+++ visualization/app/codeCharta/state/store/appSettings/showEdges/outgoing/showOutgoingEdges.reducer.ts
@@ -1,0 +1,6 @@
+import { createReducer, on } from "@ngrx/store"
+import { setState } from "../../../util/setState.reducer.factory"
+import { setShowOutgoingEdges } from "./showOutgoingEdges.actions"
+
+export const defaultShowOutgoingEdges = true
+export const showOutgoingEdges = createReducer(defaultShowOutgoingEdges, on(setShowOutgoingEdges, setState(defaultShowOutgoingEdges)))